backend/requirements.txt
Flask
flask-sqlalchemy
google-genai
